Big commit
Now there is a table for channels, one for clocks and one for symbols. The FSM doesn't contain anymore channels and clocks. There is a class called Sync which will make the synchronization with Channels. Also the expression operands have an type to search in corect table for the value. OBS is not tested yet and sync is not working